Revert "refactor: 수정된 훅을 사용하던 컴포넌트 수정"
This reverts commit 39d803123a336abffee238f8421805eb27985022.
