+++ src/components/Album/Cluster.tsx
@@ -1,4 +1,4 @@
-import { useUpdateClusterTitle } from '@/queries/album/mutations'
+import { useUpdateClusterTitle } from '@/hooks/useAlbum'
 import { ChangeEvent, CSSProperties, FocusEvent, KeyboardEvent, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import TextInput from '../common/TextInput'
